README.md
SPL Systems Programming @BGU221
my SPL course @BGU assignments with solutions, year fall 21-2022. with my friend Saker Alkoran (assignmets is for Pairs ONLY) .
This assignment was  in Social Network based on Server-Client paradigm it was for the communication between the server and the client(s) is performed using a binary communication protocol, where a registered user can follow other users and post messages.  
this assignment was in java and c++